fix(snapshot): remove unsafe pre-dispatch failure bypass
The error-class bypass canceled the mutation barrier whenever a failed
command crossed a single mutation boundary, but foreground clicks call
beginInteractionMutation() before focusing the target, so focus/activation
could already have mutated the desktop before an elementNotFound lookup
failure. That left stale snapshots resolvable (autoreview P1).

origin/main already distinguishes correctly via the tracker's
mutationStartedAt: pure pre-focus lookup failures (mutationStartedAt == nil)
cancel the durable mutation without invalidating, while any failure after a
declared mutation boundary invalidates. The live-verified 'click --on bogus'
benefit comes from that pre-existing path once the fail-open watermark
readers stop hiding snapshots -- not from the bypass. Revert the bypass,
error recorder, bridge cancel-on-pre-dispatch, and the PeekabooError/envelope
classification. Keep the fail-open readers and the honest snapshotNotAvailable
vs snapshotStale error split.
