+++ Apps/CLI/Sources/PeekabooCLI/CLI/CommanderRuntimeExecutor.swift
@@ -1,8 +1,6 @@
 import Commander
 import Foundation
 import PeekabooAutomationKit
-import PeekabooBridge
-import PeekabooFoundation
 
 /// Commands or runtime contexts that can specify a preferred capture engine.
 protocol CaptureEngineConfigurable: AnyObject {
@@ -101,25 +99,12 @@ enum CommanderRuntimeExecutor {
         }
     }
 
-    /// Everything post-command snapshot invalidation needs; split out from `CommandRuntime` so
-    /// the failure policy can be exercised by unit tests without building full runtime services.
-    struct SnapshotInvalidationDependencies {
-        let tracker: InteractionMutationTracker
-        let targets: InteractionObservationInvalidator.MutationTargets
-        let logger: Logger
-    }
-
     static func runWithImplicitSnapshotInvalidation<T>(
         using runtime: CommandRuntime,
         required: Bool,
         requiresCallerBarrier: Bool = false,
         operation: () async throws -> T
     ) async throws -> T {
-        let dependencies = SnapshotInvalidationDependencies(
-            tracker: runtime.interactionMutationTracker,
-            targets: runtime.interactionMutationTargets,
-            logger: runtime.logger
-        )
         let mutationSequenceAtStart = runtime.interactionMutationTracker.mutationSequence
         let needsCallerBarrier = required &&
             (runtime.selectedRemoteSocketPath == nil || requiresCallerBarrier)
@@ -133,7 +118,6 @@ enum CommanderRuntimeExecutor {
         } else {
             createdDurableMutation = false
         }
-        CommandFailureErrorRecorder.reset()
         let result: T
         do {
             result = try await runtime.interactionMutationTracker.withPendingDurableMutationVisible(
@@ -143,10 +127,9 @@ enum CommanderRuntimeExecutor {
             try Task.checkCancellation()
         } catch {
             _ = await self.invalidateSnapshotsAfterCommandIfNeeded(
-                dependencies: dependencies,
+                using: runtime,
                 required: required,
                 succeeded: false,
-                failure: self.underlyingCommandFailure(from: error),
                 mutationSequenceAtStart: mutationSequenceAtStart,
                 createdDurableMutation: createdDurableMutation
             )
@@ -155,7 +138,7 @@ enum CommanderRuntimeExecutor {
 
         let hadPendingMutation = required && runtime.interactionMutationTracker.mutationStartedAt != nil
         let invalidated = await invalidateSnapshotsAfterCommandIfNeeded(
-            dependencies: dependencies,
+            using: runtime,
             required: required,
             succeeded: true,
             mutationSequenceAtStart: mutationSequenceAtStart,
@@ -166,7 +149,7 @@ enum CommanderRuntimeExecutor {
         } catch {
             if hadPendingMutation {
                 _ = await self.invalidateSnapshots(
-                    dependencies: dependencies,
+                    using: runtime,
                     reason: "command cancellation",
                     through: Date(),
                     preserving: nil,
@@ -181,58 +164,27 @@ enum CommanderRuntimeExecutor {
         return result
     }
 
-    /// Commands report their real error via `handleError` and rethrow the opaque
-    /// `ExitCode.failure`; recover the recorded original so the failure can be classified.
-    private static func underlyingCommandFailure(from thrown: any Error) -> any Error {
-        let recorded = CommandFailureErrorRecorder.consume()
-        if thrown is ExitCode {
-            return recorded ?? thrown
-        }
-        return thrown
-    }
-
-    static func invalidateSnapshotsAfterCommandIfNeeded(
-        dependencies: SnapshotInvalidationDependencies,
+    private static func invalidateSnapshotsAfterCommandIfNeeded(
+        using runtime: CommandRuntime,
         required: Bool,
         succeeded: Bool,
-        failure: (any Error)? = nil,
         mutationSequenceAtStart: UInt64,
         createdDurableMutation: Bool
     ) async -> Bool {
         let completion = Date()
-        let tracker = dependencies.tracker
         guard required else { return true }
-        guard tracker.mutationStartedAt != nil else {
+        guard runtime.interactionMutationTracker.mutationStartedAt != nil else {
             guard createdDurableMutation else {
-                return !tracker.hasPendingDurableMutation
+                return !runtime.interactionMutationTracker.hasPendingDurableMutation
             }
             do {
-                try tracker.cancelDurableMutation()
+                try runtime.interactionMutationTracker.cancelDurableMutation()
                 return true
             } catch {
                 return false
             }
         }
-        if !succeeded, self.failureBypassesSnapshotInvalidation(
-            failure: failure,
-            tracker: tracker,
-            mutationSequenceAtStart: mutationSequenceAtStart
-        ) {
-            // The command failed before dispatching any desktop event, so existing snapshots stay
-            // valid. Cancel the barrier instead of advancing the watermark and hiding them all.
-            guard createdDurableMutation else { return true }
-            do {
-                try tracker.cancelDurableMutation()
-                return true
-            } catch {
-                dependencies.logger.debug(
-                    "Could not cancel the desktop mutation barrier after a pre-dispatch failure: " +
-                        error.localizedDescription
-                )
-                // Fall through to the conservative invalidation below.
-            }
-        }
-        guard let requestedCutoff = tracker.invalidationCutoff(
+        guard let requestedCutoff = runtime.interactionMutationTracker.invalidationCutoff(
             commandCompletedAt: completion,
             succeeded: succeeded
         )
@@ -240,28 +192,28 @@ enum CommanderRuntimeExecutor {
         let durableCompletion: DesktopMutationWatermarkStore.MutationCompletion?
         do {
             if createdDurableMutation,
-               tracker.mutationSequence == mutationSequenceAtStart {
-                try tracker.cancelDurableMutation()
+               runtime.interactionMutationTracker.mutationSequence == mutationSequenceAtStart {
+                try runtime.interactionMutationTracker.cancelDurableMutation()
                 durableCompletion = nil
             } else {
-                durableCompletion = try tracker.completeDurableMutation(
+                durableCompletion = try runtime.interactionMutationTracker.completeDurableMutation(
                     through: succeeded ? requestedCutoff : completion
                 )
             }
         } catch {
-            tracker.markInvalidationFailed(through: completion)
+            runtime.interactionMutationTracker.markInvalidationFailed(through: completion)
             return false
         }
         let cutoff = max(requestedCutoff, durableCompletion?.cutoff ?? requestedCutoff)
         let preservationAllowed = durableCompletion?.allowsObservationPreservation ?? true
         let preservedSnapshotID = succeeded && preservationAllowed
-            ? tracker.preservedSnapshotID
+            ? runtime.interactionMutationTracker.preservedSnapshotID
             : nil
         let preservedAt = preservedSnapshotID == nil
             ? nil
-            : tracker.preservedAt
+            : runtime.interactionMutationTracker.preservedAt
         return await self.invalidateSnapshots(
-            dependencies: dependencies,
+            using: runtime,
             reason: "command execution",
             through: cutoff,
             preserving: preservedSnapshotID,
@@ -269,46 +221,18 @@ enum CommanderRuntimeExecutor {
         )
     }
 
-    /// Whether a failed command is known not to have dispatched any desktop event, so cached
-    /// snapshots must stay resolvable. Stays conservative (returns `false`) whenever the command
-    /// crossed more than one mutation boundary (an earlier action such as a focus change may have
-    /// dispatched), a previous invalidation attempt already failed, or the error class is
-    /// ambiguous about whether an event reached the desktop (e.g. timeouts).
-    static func failureBypassesSnapshotInvalidation(
-        failure: (any Error)?,
-        tracker: InteractionMutationTracker,
-        mutationSequenceAtStart: UInt64
-    ) -> Bool {
-        guard let failure else { return false }
-        guard !tracker.hasFailedInvalidationAttempt else { return false }
-        guard tracker.mutationSequence - mutationSequenceAtStart <= 1 else { return false }
-        return self.failureLeavesDesktopUnchanged(failure)
-    }
-
-    private static func failureLeavesDesktopUnchanged(_ error: any Error) -> Bool {
-        if error is Commander.ValidationError {
-            return true
-        }
-        if let envelope = error as? PeekabooBridgeErrorEnvelope {
-            return envelope.failedBeforeDispatchingDesktopEvent
-        }
-        if let peekabooError = error as? PeekabooError {
-            return peekabooError.failedBeforeDispatchingDesktopEvent
-        }
-        return false
-    }
-
     private static func invalidateSnapshots(
-        dependencies: SnapshotInvalidationDependencies,
+        using runtime: CommandRuntime,
         reason: String,
         through cutoff: Date,
         preserving preservedSnapshotID: String?,
         preservedAt: Date?
     ) async -> Bool {
-        let isRetry = dependencies.tracker.hasFailedInvalidationAttempt
+        let targets = runtime.interactionMutationTargets
+        let isRetry = runtime.interactionMutationTracker.hasFailedInvalidationAttempt
         let invalidated = await InteractionObservationInvalidator.invalidateAfterMutation(
-            targets: dependencies.targets,
-            logger: dependencies.logger,
+            targets: targets,
+            logger: runtime.logger,
             reason: reason,
             through: cutoff,
             preserving: preservedSnapshotID,
@@ -321,8 +245,8 @@ enum CommanderRuntimeExecutor {
             return false
         }
         return await InteractionObservationInvalidator.invalidateAfterMutation(
-            targets: dependencies.targets,
-            logger: dependencies.logger,
+            targets: targets,
+            logger: runtime.logger,
             reason: "\(reason) retry",
             through: cutoff,
             preserving: preservedSnapshotID,
